# internal/appview/openapi.yaml
openapi: 3.0.3
info:
  title: HashPost AppView API
  description: HashPost AppView provides a unified API for accessing HashPost data aggregated from the PDS
  version: 1.0.0
  contact:
    name: HashPost Team
    email: team@hashpost.com
servers:
  - url: http://localhost:8081
    description: Development server
  - url: https://appview.hashpost.com
    description: Production server
paths:
...
  /api/v1/subforums:
    get:
      summary: List subforums
      description: Get a list of all subforums
      operationId: listSubforums
      tags:
        - Subforums
      parameters:
        - name: limit
          in: query
          description: Maximum number of subforums to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of subforums to skip
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of subforums
          content:
            application/json:
              schema:
                type: object
                properties:
                  subforums:
                    type: array
                    items:
                      $ref: '#/components/schemas/Subforum'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create subforum
      description: Create a new subforum (requires authentication)
      operationId: createSubforum
      tags:
        - Subforums
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - slug
                - prefix_type
              properties:
                name:
                  type: string
                  description: Name of the subforum
                  minLength: 1
                  maxLength: 255
                  example: "Technology Discussion"
                slug:
                  type: string
                  description: URL-friendly slug for the subforum (without prefix)
                  minLength: 1
                  maxLength: 255
                  pattern: '^[a-z0-9-]+$'
                  example: "technology"
                description:
                  type: string
                  description: Optional description of the subforum
                  maxLength: 1000
                  example: "A place to discuss all things tech"
                prefix_type:
                  type: string
                  description: Subforum prefix type
                  enum: ["h", "r", "t"]
                  example: "t"
      responses:
        '201':
          description: Subforum created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subforum'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/subforums/{slug}:
    get:
      summary: Get subforum by slug
      description: Get a specific subforum by its slug
      operationId: getSubforumBySlug
      tags:
        - Subforums
      parameters:
        - name: slug
          in: path
          required: true
          description: The subforum slug
          schema:
            type: string
            pattern: '^[a-z0-9-]+$'
      responses:
        '200':
          description: Subforum details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subforum'
        '404':
          description: Subforum not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/subforums/{slug}/members/{user_did}/roles:
    post:
      summary: Assign role to subforum member
      description: Assign a role to a user in a specific subforum (subforum owner/moderator only)
      operationId: assignSubforumRole
      tags:
        - Subforum Management
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          description: The subforum slug
          schema:
            type: string
            pattern: '^[a-z0-9-]+$'
        - name: user_did
          in: path
          required: true
          description: The user's DID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role_name
              properties:
                role_name:
                  type: string
                  description: Name of the role to assign
                expires_at:
                  type: string
                  format: date-time
                  description: When the role assignment expires (optional)
      responses:
        '200':
          description: Role assigned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Role assigned successfully"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permissions (must be subforum owner or moderator)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Subforum or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Revoke role from subforum member
      description: Revoke a role from a user in a specific subforum (subforum owner/moderator only)
      operationId: revokeSubforumRole
      tags:
        - Subforum Management
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          description: The subforum slug
          schema:
            type: string
            pattern: '^[a-z0-9-]+$'
        - name: user_did
          in: path
          required: true
          description: The user's DID
          schema:
            type: string
        - name: role_name
          in: query
          required: true
          description: Name of the role to revoke
          schema:
            type: string
      responses:
        '200':
          description: Role revoked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Role revoked successfully"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permissions (must be subforum owner or moderator)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Subforum, user, or role not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # User Self-Management Endpoints
...
  /api/v1/posts/{id}/moderation:
    get:
      summary: Get post moderation state
      description: Get moderation flags (pinned, locked, removed) for a post
      operationId: getPostModerationState
      tags:
        - Posts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Post ID
      responses:
        '200':
          description: Post moderation state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModerationState'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Subscription endpoints
...
components:
  schemas:
...
    Subforum:
      type: object
      required:
        - id
        - name
        - slug
        - created_by
        - created_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the subforum
        name:
          type: string
          maxLength: 100
          description: Display name of the subforum
        slug:
          type: string
          maxLength: 50
          pattern: '^[a-z0-9-]+$'
          description: URL-friendly identifier for the subforum
        description:
          type: string
          maxLength: 500
          description: Description of the subforum
        created_by:
          type: string
          format: uuid
          description: ID of the user who created the subforum
        created_at:
          type: string
          format: date-time
          description: When the subforum was created
        updated_at:
          type: string
          format: date-time
          description: When the subforum was last updated
        subscriber_count:
          type: integer
          minimum: 0
          description: Number of subscribers to the subforum
        post_count:
          type: integer
          minimum: 0
          description: Number of posts in the subforum
        comment_count:
          type: integer
          minimum: 0
          description: Number of comments in the subforum
        prefix_type:
          type: string
          enum: ["h", "r", "t"]
          description: Subforum prefix type (h=HashPost, r=regional, t=topical)
...
    ModerationState:
      type: object
      required:
        - is_pinned
        - is_locked
        - is_removed
      properties:
        is_pinned:
          type: boolean
          description: Whether the post is pinned
        is_locked:
          type: boolean
          description: Whether the post is locked
        is_removed:
          type: boolean
          description: Whether the post is removed by moderation
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details

    # RBAC Schemas
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from PDS authentication
...